Move test promotion factory to Behat-related services
