+++ src/Sylius/Bundle/CoreBundle/Resources/config/test_services.xml
@@ -47,9 +47,6 @@
             <argument>%locale%</argument>
         </service>
 
-        <service id="sylius.test.factory.promotion" class="Sylius\Component\Core\Test\Factory\TestPromotionFactory">
-            <argument type="service" id="sylius.factory.promotion" />
-        </service>
     </services>
 
 </container>
